feat: add finalize command
