+++ packages/cli/src/options.ts
@@ -1,0 +1,6 @@
+import { Option } from "commander";
+
+export const parallelNonce = new Option(
+  "--parallel-nonce <string>",
+  "A unique ID for this parallel build",
+).env("ARGOS_PARALLEL_NONCE");
