+++ packages/core/src/auth.test.ts
@@ -1,80 +1,6 @@
-import { describe, it, expect, beforeAll } from "vitest";
+import { describe, it, expect } from "vitest";
 
-import { setupMockServer } from "../mocks/server";
-import {
-  ArgosApiClient,
-  createArgosLegacyAPIClient,
-  getAuthToken,
-} from "./api-client";
-
-setupMockServer();
-
-let apiClient: ArgosApiClient;
-
-describe("#createArgosLegacyAPIClient", () => {
-  beforeAll(() => {
-    apiClient = createArgosLegacyAPIClient({
-      baseUrl: "https://api.argos-ci.dev",
-      bearerToken: "Bearer 92d832e0d22ab113c8979d73a87a11130eaa24a9",
-    });
-  });
-
-  describe("#createBuild", () => {
-    it("creates build", async () => {
-      const result = await apiClient.createBuild({
-        commit: "f16f980bd17cccfa93a1ae7766727e67950773d0",
-        screenshotKeys: ["123", "456"],
-        pwTraceKeys: [],
-      });
-      expect(result).toEqual({
-        build: {
-          id: "123",
-          url: "https://app.argos-ci.dev/builds/123",
-        },
-        screenshots: [
-          {
-            key: "123",
-            putUrl: "https://api.s3.dev/upload/123",
-          },
-          {
-            key: "456",
-            putUrl: "https://api.s3.dev/upload/456",
-          },
-        ],
-      });
-    });
-  });
-
-  describe("#updateBuild", () => {
-    it("updates build", async () => {
-      const result = await apiClient.updateBuild({
-        buildId: "123",
-        screenshots: [
-          {
-            key: "123",
-            name: "screenshot 1",
-            metadata: null,
-            pwTraceKey: null,
-            threshold: null,
-          },
-          {
-            key: "456",
-            name: "screenshot 2",
-            metadata: null,
-            pwTraceKey: null,
-            threshold: null,
-          },
-        ],
-      });
-      expect(result).toEqual({
-        build: {
-          id: "123",
-          url: "https://app.argos-ci.dev/builds/123",
-        },
-      });
-    });
-  });
-});
+import { getAuthToken } from "./auth";
 
 describe("#getAuthToken", () => {
   describe("without CI", () => {
