+++ packages/core/src/auth.ts
