+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./upload";
+export * from "./finalize";
 export * from "./config";
